[*] class recomposition
